==> code/02_population_interpolation.ipynb <==
mortality_url = 'https://raw.githubusercontent.com/mariobecerra/mda_project/main/data/mortality_data_2000-2019.csv'
population_url = 'https://raw.githubusercontent.com/mariobecerra/mda_project/main/data/Population_par_commune.xlsx'

import pandas as pd
import numpy as np

#################                weekly mortality data divided by municipality (2000-2019)
week_mor_mun = pd.read_csv(mortality_url)
week_mor_mun['WEEK'] = week_mor_mun['YEAR_WEEK'].str.slice(start=6).astype('int') # get week number
week_mor_mun.rename(columns = {'N_MASK':'N_DEATHS'}, inplace = True)
del week_mor_mun['YEAR_WEEK']
week_mor_mun = week_mor_mun[week_mor_mun.COD != 'external'] # do not need data for external causes of death
week_mor_mun = week_mor_mun[week_mor_mun.ARRON != 58000] # data for Arrondissement de La Louvière is only available for 2019 and therefore has been removed from dataset
week_mor_mun.head()

##################               combine data sets from different excel sheets

# data with yearly population data divided by municipality
xls = pd.ExcelFile(population_url)

# create base dataframe
cols = ['municipality_code', 'municipality_name', 'male', 'female', 'total']
year_pop_mun = pd.DataFrame(columns=cols)
years = []
for i in range(21): # need poulation data for 21 years (years on both ends)
    df = pd.read_excel(xls, 'Population en '+str(2000+i), header=None, skiprows=4, nrows=646) # 2000-2019
    df.columns = cols
    length_of_df = df.loc[df['municipality_name'] == 'Viroinval'].index.tolist()[0] + 1 # get the last line of table
    years = years+[2000+i]*length_of_df
    year_pop_mun = pd.concat([year_pop_mun, df[:length_of_df]] , axis=0, ignore_index=True)
year_pop_mun['year'] = years

# convert population columns to int
convert_dict = {'male': int, 'female': int, 'total': int, 'municipality_code':int}
year_pop_mun = year_pop_mun.astype(convert_dict)

year_pop_mun

##################                assign population to weekly death rate data (at week 1)
existing_arrons = week_mor_mun['ARRON'].value_counts().index
week_mor_mun['population_t'] = np.nan
exceptions = []
for year in range(21):
    for arron in existing_arrons:
        value = year_pop_mun[(year_pop_mun['year'] == (2000+year)) & (year_pop_mun['municipality_code'] == arron)]['total'].tolist()
        if len(value) == 1: # check if population value exists in year_pop_mun
            condition = (week_mor_mun['ARRON'] == arron) & (week_mor_mun['WEEK'] == 1) & (week_mor_mun['YEAR'] == (2000+year))
            if (condition).any(): # check if a row with week 1 exists in week_mor_mun
                week_mor_mun.loc[condition, 'population_t'] = value[0]
            else:
                dummy_df = pd.DataFrame({'YEAR': [2000+year], 'COD': ['natural'], 'ARRON': [arron], 'N_DEATHS': [0], 'WEEK':[1], 'population_t':  [value[0]]})
                week_mor_mun = pd.concat([week_mor_mun, dummy_df] , axis=0, ignore_index=True)
        else:
            exceptions.append((year, arron))

exceptions

week_mor_mun['ARRON'].value_counts()

# get number of weeks for each year
no_weeks = [week_mor_mun[(week_mor_mun['YEAR'] == (2000+i))]['WEEK'].max() for i in range(20)] # weeks for each year
for year in range(20): # for each year
    for arron in existing_arrons: # for each arron
        if arron == 54000 and (year == 18 or year == 19): # 54000 population not available for 2019
            continue
        condition = (week_mor_mun['ARRON'] == arron) & (week_mor_mun['WEEK'] == 1)
        prev_year_pop = week_mor_mun.loc[condition & (week_mor_mun['YEAR'] == (2000+year)), 'population_t'].tolist()[0]
        next_year_pop = week_mor_mun.loc[condition & (week_mor_mun['YEAR'] == (2000+year+1)), 'population_t'].tolist()[0]
        inter_values = np.linspace(prev_year_pop, next_year_pop, no_weeks[year], endpoint=False, dtype=int)[1:]
        for ix, k in enumerate(inter_values):
            week_mor_mun.loc[(week_mor_mun['ARRON'] == arron) & (week_mor_mun['YEAR'] == (2000+year)) & (week_mor_mun['WEEK'] == ix + 2),'population_t'] = k

print(arron, year)

week_mor_mun.head(5)

week_mor_mun.to_csv('../out/weekly_mortality_with_linear_interpolated_population_data.csv', index=False)

# Mortality dataframe for modeling

df = pd.read_csv(mortality_url)
df.rename(columns = {'N_MASK':'N_DEATHS'}, inplace = True)
df['WEEK'] = df['YEAR_WEEK'].str.slice(start=6).astype('int') # get week number
df = df[df.COD != 'external'] # do not need data for external causes of death
df = df[df.ARRON != 58000] # data for Arrondissement de La Louvière is only available for 2019 and therefore has been removed from dataset
df = df.sort_values(by = ['YEAR_WEEK', 'ARRON'])
df.to_csv('../out/mortality.csv', index=False)
