==> code/03_merge_mortality_population.ipynb <==
import pandas as pd
import numpy as np

'LOAD HEAT WAVE DATA'
heat_wave = pd.read_csv("../out/mortality_heat_wave.csv", sep = ",")
print(heat_wave)

'LOAD DEATH DATA WITH TOTAL WEEKLY POPULATION BY ARRONDISSEMENT (AS OBTAINED BY LINEAR INTERPOLATION)'
mort_interp = pd.read_csv("../out/weekly_mortality_with_linear_interpolated_population_data.csv",
                         sep = ",")
print(mort_interp)
mort_interp.info()

mort_interp[mort_interp['ARRON']=='21000']


heat_wave.head()

'FILTER HEAT_WAVE DATA TO GET ONLY NATURAL DEATHS'
heat_wave = heat_wave[heat_wave['COD'] == "natural"]

'CONVERT YEAR_WEEK TO STRING'
heat_wave = heat_wave.astype({"YEAR_WEEK": str})

'CREATING YEAR AND WEEK VARIABLES'
heat_wave['YEAR'] = heat_wave['YEAR_WEEK'].astype(str).str[:4]
heat_wave['YEAR'] = heat_wave['YEAR'].astype(int)

heat_wave['WEEK'] = heat_wave['YEAR_WEEK'].astype(str).str[6:8]
heat_wave['WEEK'] = heat_wave['WEEK'].astype(int)

print(heat_wave)
heat_wave.info()

heat_wave[heat_wave['ARRON']==21000]

'MERGING HEAT_WAVE AND INTERPOLATION DATASETS'

merged_data = pd.merge(mort_interp, heat_wave, how = 'right', left_on=['YEAR', 'ARRON', 'WEEK'], right_on = ['YEAR', 'ARRON', 'WEEK'])
merged_data[merged_data['N_DEATHS'].isnull()& merged_data['N_MASK'].notnull()]
#no obsevation with null N_MASK/COD_y and not null N_DEATHS/COD_x, so drop N_DEATHS/COD_x
merged_data=merged_data.drop(columns=['COD_x','N_DEATHS','NIS_Code','Unnamed: 0'])
merged_data.head()
#merged_data.info()

'SAVING MERGED DATA TO OUT FILE'
merged_data.to_csv('../out/merged.csv', encoding = 'utf-8-sig')
