==> code/04_create_heatwave_frequency_and_intensity_dataframe_1980-2020.ipynb <==
URLs of data

arrondissements_url = 'https://raw.githubusercontent.com/mariobecerra/mda_project/main/data/arrondissements_coords.csv'
temperature_url = 'https://raw.githubusercontent.com/mariobecerra/mda_project/main/data/T_Belgium_1980-2020.csv' 

Read data

import pandas as pd
import numpy as np
import datetime

temperature_1980_2020 = pd.read_csv(temperature_url, sep=";")
temperature_1980_2020.head(5) 

arrondissements_data = pd.read_csv(arrondissements_url)
arrondissements_data.head(5)

Find the closest temperature grid to each arrondissement

grids_temp = temperature_1980_2020.drop_duplicates(subset = ['GRID_NO', 'LATITUDE', 'LONGITUDE'])
grids_temp

# Not the most efficient, but it's a small dataset so it's okay
distances = pd.DataFrame()

for i in range(arrondissements_data.shape[0]):
  diff_lat = grids_temp['LATITUDE'] - arrondissements_data.loc[i].at['lat']
  diff_lon = grids_temp['LONGITUDE'] - arrondissements_data.loc[i].at['lon']
  distances_i = np.power((diff_lat), 2) + np.power((diff_lon), 2)

  temp = pd.DataFrame(
        {
            'NIS_Code': arrondissements_data.loc[i].at['NIS_Code'],
            'GRID_NO': grids_temp.GRID_NO,
            'Dist': distances_i
        }
    )
  distances = pd.concat([distances, temp])

distances

Find the temperature grid closest to each arrondissement

arron_grid_mapping = distances.groupby('NIS_Code').apply(lambda x: x[x['Dist'] == x['Dist'].min()])[['NIS_Code', 'GRID_NO']].reset_index(drop = True)

arron_grid_mapping.head(8)

Join arrondissement and temperature datasets to get temperature per day in each arrondissement.

join_1 = pd.merge(arron_grid_mapping, temperature_1980_2020, how = "inner")
join_1

arron_temp_data = pd.merge(join_1, arrondissements_data, how = 'left')[['DAY', 'Nom_arrondissement', 'NIS_Code', 'GRID_NO', 'TEMPERATURE_MAX', 'TEMPERATURE_MIN', 'TEMPERATURE_AVG']].sort_values(by=['Nom_arrondissement', 'NIS_Code', 'DAY']).assign(YEAR = lambda x: np.floor(x.DAY/10000))
arron_temp_data

Get percentiles to define heatwave

def q95(x):
    return x.quantile(0.95)

temp_percetiles_year_arron = arron_temp_data[['YEAR', 'NIS_Code', 'Nom_arrondissement', 'TEMPERATURE_MAX']].groupby(['YEAR', 'NIS_Code', 'Nom_arrondissement']).agg(q95).stack(level=0).reset_index().rename(columns={0:"p95"})
temp_percetiles_year_arron

Get temperatures of five consecutive days

temp_lead_1 = arron_temp_data.groupby(['Nom_arrondissement', 'NIS_Code'])['TEMPERATURE_MAX'].shift(-1)

temp_lead_2 = arron_temp_data.groupby(['Nom_arrondissement', 'NIS_Code'])['TEMPERATURE_MAX'].shift(-2)

temp_lead_3 = arron_temp_data.groupby(['Nom_arrondissement', 'NIS_Code'])['TEMPERATURE_MAX'].shift(-3)

temp_lead_4 = arron_temp_data.groupby(['Nom_arrondissement', 'NIS_Code'])['TEMPERATURE_MAX'].shift(-4)

arron_temp_data['temp_lead_1'] = temp_lead_1
arron_temp_data['temp_lead_2'] = temp_lead_2
arron_temp_data['temp_lead_3'] = temp_lead_3
arron_temp_data['temp_lead_4'] = temp_lead_4
arron_temp_data

Create boolean variable if there are 5 consecutive days in which temperature was higher than the 95-th percentile of temperature each year and each arrondissement

join_percentiles = pd.merge(arron_temp_data, temp_percetiles_year_arron[['YEAR', 'NIS_Code', 'Nom_arrondissement', 'p95']], how = "left")
join_percentiles

nan_values = join_percentiles[join_percentiles.isna().any(axis=1)]

print (nan_values)

bool0 = (join_percentiles['TEMPERATURE_MAX'].to_numpy() >= join_percentiles['p95'].to_numpy()).astype(int)
bool1 = (join_percentiles['temp_lead_1'].to_numpy() >= join_percentiles['p95'].to_numpy()).astype(int)
bool2 = (join_percentiles['temp_lead_2'].to_numpy() >= join_percentiles['p95'].to_numpy()).astype(int)
bool3 = (join_percentiles['temp_lead_3'].to_numpy() >= join_percentiles['p95'].to_numpy()).astype(int)
bool4 = (join_percentiles['temp_lead_4'].to_numpy() >= join_percentiles['p95'].to_numpy()).astype(int)
heatwave_boolean = bool0 * bool1 * bool2 * bool3 * bool4
heatwave_boolean

# Not the most efficient, but it works
iso_weeks = ['1990-001']*join_percentiles.shape[0] # Prefill list with correct size
for i in range(len(iso_weeks)):
  day_i = join_percentiles.loc[i].at['DAY']
  iso_week_i = datetime.datetime.strptime(str(day_i), '%Y%m%d').isocalendar()
  iso_weeks[i] = str(iso_week_i[0]) + '-' + str(iso_week_i[1]).zfill(3)

join_percentiles.columns

heat_wave_day_def = join_percentiles[['NIS_Code', 'Nom_arrondissement','YEAR', 'TEMPERATURE_MAX', 'TEMPERATURE_AVG', 'p95']].copy()
heat_wave_day_def['heatwave_boolean'] = heatwave_boolean
heat_wave_day_def['YEAR'] = heat_wave_day_def['YEAR'].astype(int)
heat_wave_day_def

heat_wave_day_def['tempdiff_max_p95'] = heat_wave_day_def['TEMPERATURE_MAX'] - heat_wave_day_def['p95'] 
heat_wave_day_def.head()

Calculate the heat waves frequency (the number of episodes per year) and the intensity (average T above the local threshold during heat waves) per arrondissement

arrons_iter = heat_wave_day_def.NIS_Code.unique().tolist()
start_yr = int(heat_wave_day_def.YEAR.min())
stop_yr = int(heat_wave_day_def.YEAR.max())
years_iter = [i for i in range(start_yr, stop_yr+1)]
hw_count_df_cols = ['arron'] + years_iter
hw_count_df = pd.DataFrame(columns=hw_count_df_cols)
hw_int_df = pd.DataFrame(columns=hw_count_df_cols)
hw_count_df['arron'] = arrons_iter
hw_int_df['arron'] = arrons_iter

heat_wave_intensity = heat_wave_day_def[heat_wave_day_def['heatwave_boolean'] == 1].copy()
summarized_temp_for_hw_intensity = heat_wave_intensity.groupby(['NIS_Code', 'YEAR'])[['tempdiff_max_p95']].mean().reset_index()

summarized_counts_for_hw = heat_wave_day_def.groupby(['NIS_Code', 'YEAR'])[['heatwave_boolean']].sum().reset_index()

for arr in arrons_iter:
  for yrs in years_iter:
    ix = hw_count_df.index[hw_count_df['arron'] == arr].tolist()[0]
    val = summarized_counts_for_hw[(summarized_counts_for_hw.NIS_Code == arr) & (summarized_counts_for_hw.YEAR == yrs)]['heatwave_boolean'].tolist()[0]
    try:
      val_i = summarized_temp_for_hw_intensity[(summarized_temp_for_hw_intensity.NIS_Code == arr) & (summarized_temp_for_hw_intensity.YEAR == yrs)]['tempdiff_max_p95'].tolist()[0]
    except:
      val_i = 0
    hw_count_df.loc[ix , yrs] = val
    hw_int_df.loc[ix, yrs] = val_i
    

hw_int_df.head(5)

hw_count_df.head(5)

save the file with heat wave intensities and frequencies

hw_int_df.to_csv('../out/hw_int_df.csv', encoding = 'utf-8-sig', index=False) 
hw_count_df.to_csv('../out/hw_freq_df.csv', encoding = 'utf-8-sig', index=False)
