==> code/07_prepare_data_for_modeling.ipynb <==
# Dataframe for Meta-Analysis

import pandas as pd
import datetime
import numpy as np
import patsy 

import warnings
warnings.filterwarnings('ignore')

# Load arron-temp data
arron_temp = pd.read_csv("https://raw.githubusercontent.com/mariobecerra/mda_project/main/out/arron_temp_data.csv", sep = ",")

# Load mortality data
mort = pd.read_csv("https://raw.githubusercontent.com/mariobecerra/mda_project/main/out/mortality.csv", sep = ",")

mort

arron_temp = arron_temp.drop(['Unnamed: 0', 'GRID_NO'], axis = 1)

arron_temp

regions = arron_temp['Nom_arrondissement'].unique()
datalist = {}

for region in regions:
    datalist[region] = arron_temp[arron_temp['Nom_arrondissement'] == region]
    datalist[region] = datalist[region].rename(columns={"DAY": "DATE"})
    datalist[region]['tmean05'] = datalist[region]['TEMPERATURE_AVG'].rolling(window=6).mean()
    datalist[region]['YEAR'] = [int((x//10000)) for x in datalist[region]['DATE']]
    datalist[region]['MONTH'] = [int((x//100)%100) for x in datalist[region]['DATE']]
    datalist[region]['DAY'] = [int(x%100) for x in datalist[region]['DATE']]
    datalist[region]['WEEK'] = [int(datetime.date(x//10000, (x//100)%100, x%100).strftime("%V")) for x in datalist[region]['DATE']]
    datalist[region]['YDAY'] = [int(datetime.date(x//10000, (x//100)%100, x%100).strftime('%j')) for x in datalist[region]['DATE']]
    datalist[region]['DOW'] = [int(datetime.date(x//10000, (x//100)%100, x%100).weekday()) for x in datalist[region]['DATE']]
    datalist[region]['TIME'] = list(range(1, len(datalist[region]['DATE']) + 1))
    datalist[region] = datalist[region][['DATE', 'YEAR', 'MONTH', 'DAY', 'WEEK', 'YDAY', 'DOW', 'TIME', 'Nom_arrondissement', 'NIS_Code', 'TEMPERATURE_MIN', 'TEMPERATURE_MAX', 'TEMPERATURE_AVG', 'tmean05']]
    datalist[region]['N_DEATHS'] = -1
    datalist[region].reset_index(drop = True)
    
    n_deaths=[]
    for i in range(0, len(datalist[region])):
        deaths = mort[(mort['YEAR']==datalist[region].iloc[i]['YEAR']) & 
                      (mort['WEEK']==datalist[region].iloc[i]['WEEK']) & 
                      (mort['ARRON']==datalist[region].iloc[i]['NIS_Code'])]['N_DEATHS']
        if len(deaths) ==0:
            n_deaths.append(np.NAN)
        else:
            n_deaths.append(deaths.iloc[0])
    datalist[region]['N_DEATHS'] = [round(d/7, 2) for d in n_deaths]

datalist['Dinant']

# saving meta-analysis dataset to /out
df = pd.DataFrame(columns = (datalist['Dinant']).columns)
for region in regions:
    df = pd.concat([df, datalist[region]])

df.to_csv('../out/meta_analysis_data.csv', encoding = 'utf-8-sig') 

df

# Model

m = len(datalist)
m

# TEMPERATURE RANGES (FOR LAG 0-5)
min_tmean05 = pd.DataFrame([{k:(v['tmean05']).min() for k,v in datalist.items()}], index =["MIN"])
max_tmean05 = pd.DataFrame([{k:(v['tmean05']).max() for k,v in datalist.items()}], index = ["MAX"])
ranges = pd.concat([min_tmean05, max_tmean05], axis = 0).transpose()
ranges

# DEFINE THE AVERAGE RANGE, CENTERING POINT, DEGREE AND TYPE OF THE SPLINE
# (THESE PARAMETERS CAN BE CHANGED BY THE USER FOR ADDITIONAL ANALYSES)
cen = 17
bounds = ranges.apply(np.mean)
degree = 2
type = "bs"
df = 6

# DEFINE THE KNOTS AT TEMPERATURE CORRESPONDING TO AVERAGE PERCENTILES
knotperc = [5, 35, 65, 95]
knots = pd.DataFrame(columns = ["5%", "35%", "65%", "95%"], index = regions)
for region in regions:
    perc = np.nanpercentile(datalist[region]["tmean05"], knotperc)
    knots.loc[region] = [perc[0], perc[1], perc[2], perc[3]]
knots
  

btmean05 = patsy.bs(x=datalist['Dinant']['tmean05'], knots=knots.loc['Dinant'], degree=degree, lower_bound=bounds[0], upper_bound=bounds[1])
btmean05
